+++ src/app.d.ts
@@ -9,3 +9,15 @@ declare namespace App {
 	// interface Session {}
 	// interface Stuff {}
 }
+
+
+declare module "*&picture" {
+	/**
+	 * actual types
+	 * taken from https://github.com/JonasKruckenberg/imagetools/issues/160#issuecomment-1009292026
+	 * - code https://github.com/JonasKruckenberg/imagetools/blob/main/packages/core/src/output-formats.ts
+	 * - docs https://github.com/JonasKruckenberg/imagetools/blob/main/docs/guide/getting-started.md#metadata
+	 */
+	const out;
+	export default out;
+  }
