+++ src/lib/components/Picture.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   // See: https://github.com/JonasKruckenberg/imagetools/blob/main/docs/directives.md#picture
-  export let data;
+  import type { Picture } from 'vite-imagetools';
+  export let data: Picture;
   export let alt: string;
 </script>
