+++ src/routes/docs/+layout.server.ts
@@ -1,4 +1,4 @@
-import type { PageServerLoad } from './$types';
+import type { LayoutServerLoad } from './$types';
 
 import { index_content } from './documentation.server';
 
@@ -7,4 +7,4 @@ import { index_content } from './documentation.server';
 // If you can no longer prerender the docs, then you are going to have to move the load functions here to a prerendered server route like before and fetch them here.
 export const prerender = true;
 
-export const load: PageServerLoad = () => ({ tree: index_content() });
+export const load: LayoutServerLoad = () => ({ tree: index_content() });
